* process.c: suppress warning for signed and unsigned type   inconsistency.
* ext/psych/parser.c: ditto.

* ext/sdbm/_sdbm.c: ditto.

* ext/syck/rubyext.c: ditto.

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@27864 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ ext/syck/rubyext.c
@@ -275,7 +275,7 @@ mktime_do(struct mktime_arg *arg)
         char *end = ptr + 1;
         char *p = end;
         while ( isdigit( *end ) ) end++;
-        if (end - p < sizeof(padded)) {
+        if (end - p < (int)sizeof(padded)) {
             MEMCPY(padded, ptr + 1, char, end - (ptr + 1));
             p = padded;
         }
